reorganized code - md
