+++ tests/unitary/with_extras/aqua/test_deployment.py
@@ -17,14 +17,13 @@ import pytest
 import yaml
 from parameterized import parameterized
 
-import ads.aqua.deployment
+import ads.aqua.modeldeployment.deployment
 import ads.config
 from ads.aqua.common.errors import AquaRuntimeError
-from ads.aqua.deployment import (
+from ads.aqua.modeldeployment import AquaDeploymentApp, MDInferenceResponse
+from ads.aqua.modeldeployment.entities import (
     AquaDeployment,
-    AquaDeploymentApp,
     AquaDeploymentDetail,
-    MDInferenceResponse,
     ModelParams,
 )
 from ads.model.datascience_model import DataScienceModel
@@ -170,7 +169,7 @@ class TestAquaDeployment(unittest.TestCase):
         os.environ["PROJECT_COMPARTMENT_OCID"] = TestDataset.USER_COMPARTMENT_ID
         reload(ads.config)
         reload(ads.aqua)
-        reload(ads.aqua.deployment)
+        reload(ads.aqua.modeldeployment.deployment)
 
     @classmethod
     def tearDownClass(cls):
@@ -180,7 +179,7 @@ class TestAquaDeployment(unittest.TestCase):
         os.environ.pop("PROJECT_COMPARTMENT_OCID", None)
         reload(ads.config)
         reload(ads.aqua)
-        reload(ads.aqua.deployment)
+        reload(ads.aqua.modeldeployment.deployment)
 
     def test_list_deployments(self):
         """Tests the list method in the AquaDeploymentApp class."""
@@ -203,7 +202,7 @@ class TestAquaDeployment(unittest.TestCase):
                 expected_attributes
             ), "Attributes mismatch"
 
-    @patch("ads.aqua.deployment.get_resource_name")
+    @patch("ads.aqua.modeldeployment.deployment.get_resource_name")
     def test_get_deployment(self, mock_get_resource_name):
         """Tests the get method in the AquaDeploymentApp class."""
 
@@ -256,7 +255,7 @@ class TestAquaDeployment(unittest.TestCase):
 
             self.app.get(model_deployment_id=TestDataset.MODEL_DEPLOYMENT_ID)
 
-    @patch("ads.aqua.deployment.load_config")
+    @patch("ads.aqua.modeldeployment.deployment.load_config")
     def test_get_deployment_config(self, mock_load_config):
         """Test for fetching config details for a given deployment."""
 
@@ -277,7 +276,7 @@ class TestAquaDeployment(unittest.TestCase):
 
     @patch("ads.aqua.deployment.get_container_config")
     @patch("ads.aqua.model.AquaModelApp.create")
-    @patch("ads.aqua.deployment.get_container_image")
+    @patch("ads.aqua.modeldeployment.deployment.get_container_image")
     @patch("ads.model.deployment.model_deployment.ModelDeployment.deploy")
     def test_create_deployment_for_foundation_model(
         self,
@@ -342,7 +341,7 @@ class TestAquaDeployment(unittest.TestCase):
 
     @patch("ads.aqua.deployment.get_container_config")
     @patch("ads.aqua.model.AquaModelApp.create")
-    @patch("ads.aqua.deployment.get_container_image")
+    @patch("ads.aqua.modeldeployment.deployment.get_container_image")
     @patch("ads.model.deployment.model_deployment.ModelDeployment.deploy")
     def test_create_deployment_for_fine_tuned_model(
         self,
